read: Display control and non-control characters
If the character that is read() is a control character, it is directly
displayed in %d format. On the other hand, if it is a non control
character, it is printed in both integer and character format.

In this way, we get to see how various keypresses are translated as
bytes when they are read. Escape sequence is inputs 3 or 4 byte to the
terminal and starts with 27, '[' followed by one or two characters.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>