flags: Turned off XOFF, XON, EXTEN and CRNL flags
The XOFF and XON flags were causing the program to pause and
resume respectively. Now, ctrl-S and ctrl-Q will be read as
19 byte and 17 byte respectively.

The EXTEN flag was causing the terminal to pause until the
next character is typed. Now ctrl-V can be read as 22 byte.

This also fixes ctrl-O problem in MacOS. ctrl-O will be read as
15 bytes.

The ctrl-M  is read as 10 bytes while it is expected to be read as
13 byte. By turning of CRNL ctrl-M can be read as 13 byte.

fixes: #8

Signed-off-by: satyam kumar <satyamvats5@gmail.com>
Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>