flags: Turn off all output and other flags.
OPOST: The terminal does a translation on the output side,
it translates each new line ("\n") into cariage return
followed by new line ("\r\n"). We are going to stop this
output processing.

BRKINT: It cause a SIGINT signal to be sent to the program,
just like pressing ctrl-C.

INPCK: It enables parity checking , althouh this does not
any visible effects on these modern days terminals.

ISTRIP: It causes each 8th bit of each input byte to be
stripped.

CS8: It is not a flag, it is a bit mask with multiple
bits, which we set using the bitwise-OR (|) operator
unlike all the flags we are turning off.

fixes: https://github.com/BytesClub/chalk/issues/10

Signed-off-by: satyam kumar <satyamvats5@gmail.com>
Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>