Draw: Hiding the cursor while repainting
To overcome the flicker effect we are hiding the cursor while
repainting.
We are using command "[?25l" and "[?25h" preceedded by escape
sequence for hiding and showing the cursor respectively.
We are formating each line seperatey by using command "[K"
(which is equivalent to "[0K"), preceeded by escape sequence.
"[k" formats the part of line which is right to the cursor.

fixes: #23 (https://github.com/BytesClub/chalk/issues/23)

Signed-off-by: SATYAM KUMAR <satyamvats5@gmail.com>
Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>