move: Move the cursor
We are moving the cursor by mapping the inputs provided by
the arrow keypresses to moveCursor() function.
As we know that arrow key returns a sequence of 3-4 bytes
which includes escape sequence followed by some characters,
we will use this to detect whether the key pressesd is an arrow
key or not.
Arrow key returns some characters precceded by escape sequence.
We will first check whether the input character is an escape sequence,
if it is than we will read two other characters and check whether it is
an arrow key or not.
Arrow_Keys returns '\x1b', '[', followed by an 'A', 'B', 'C', or 'D' .
Page_Up and Page_down returns "<esc>[5~" and "<esc>[6~" respectively.
Home key could be sent as "<esc>[1~", "<esc>[7~", "<esc>[H", or "<esc>OH"
Whereas End_Key could be send as "<esc>[4~", "<esc>[8~", "<esc>[F", or
"<esc>OF".
We are also mapping Delete_Key although we are not doing anything with it for now.
Delete_Key sends "<esc>[3~"

fixes: #24 ( https://github.com/BytesClub/chalk/issues/24 )

Signed-off-by: SATYAM KUMAR <satyamvats5@gmail.com>
Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>