display: Display the contents
We are creating a data type to store the elements of a row and size
of that row, here we are creating a data type 'srow' using typedef.
We are using two variables named linecap and line len of data type size_t
and ssize_t respectively. size_t and ssize_t comes under header file
<sys/types.h> .
size_t is of unsigned int type (It is an alias which stands for sizeof type
declared as:- typedef unsigned int size_t ). Whenever there is need to store
size_of one variable to other we use size_t type.
ssize_t is similar to size_t but it is of int type instead of unsigned int type.
(declared as:- typedef int ssize_t), it is used when there is chance of memory
allocation failure.

We are copying the elements of rows to our global variable E.row using function
appendrow() and openeditor().
We are taking name of a file through Command Line Argument from the user.
If a valid file name is passed by  the user than, the details of that file is
shown on our screen and if not then welcome screen is shown on the screen.

fixes: #26 (https://github.com/BytesClub/chalk/issues/26)

Signed-off-by: satyam kumar <satyamvats5@gmail.com>
Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>