docs: rename install to HOWTOs
Since the install.rst was not actually a guide to install chalk to the
system, but an instruction to build the project from source, it was
renamed accordingly. Added the banner image of chalk in the front page
and updated git add . to git add <files> and added -v flag in the git
commit command to view the diff while making the commit.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>